feat: Add encode and decode messages in decoder.js
This commit adds the functionality to encode and decode messages in the decoder.js file. It includes the addition of the `encodeMessage` and `decodeMessage` variables, which respectively encode and decode the message "Code by ybtuti". This enhancement improves the functionality of the project by providing the capability to encode and decode messages.